+++ Examples/BasicAuth/BasicAuth/AppDelegate.h
@@ -1,0 +1,19 @@
+//
+//  AppDelegate.h
+//  BasicAuth
+//
+//  Created by Nick Lockwood on 17/05/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@class ViewController;
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@property (strong, nonatomic) ViewController *viewController;
+
+@end