+++ Examples/BasicAuth/BasicAuth/ViewController.h
@@ -1,0 +1,15 @@
+//
+//  ViewController.h
+//  BasicAuth
+//
+//  Created by Nick Lockwood on 17/05/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface ViewController : UIViewController
+
+@property (nonatomic, strong) IBOutlet UIImageView *imageView;
+
+@end